requirements.txt
mautrix>=0.15.5,<0.19
aiohttp>=3,<4
yarl>=1,<2
asyncpg>=0.20,<0.26
ruamel.yaml>=0.15.35,<0.18
attrs>=18.1.0
jinja2>=2,<4
